Adding ingredients and steps
The last block of code creates an instance of a new recipe where it takes the previous ingredients saved inn the array as well as the number of steps that was also added in an array and saved in the new Recipe. It sets the factor scale to 1 (default) while showing the user the recipe has been added.