Implement recipe scaling with user input
This commit I have added a function ScaleRecipe which allows users to interactively scale a recipe. The function checks if a recipe is loaded and if not informs the user. If it is, it will present the user with 3 options to scale the recipe by. It validates user input and applies the chosen scale factor to the recipes ingredients.